+++ src/test/java/StatsServiceTest.java
@@ -7,11 +7,10 @@ public class StatsServiceTest {
 
     @Test
     public void shouldCalculateMinSales() {
-
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 9;
-        int actual = service.calculateMinSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 9;
+        long actual = service.calculateMinSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -19,9 +18,9 @@ public class StatsServiceTest {
     @Test
     public void shouldCalculateSumOfAllSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 180;
-        int actual = service.calculateSumOfAllSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 180;
+        long actual = service.calculateSumOfAllSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -29,9 +28,9 @@ public class StatsServiceTest {
     @Test
     public void shouldCalculateAverageSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 15;
-        int actual = service.calculateAverageSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 15;
+        long actual = service.calculateAverageSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -39,9 +38,9 @@ public class StatsServiceTest {
     @Test
     public void shouldCalculateMaximumSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 8;
-        int actual = service.calculateMaximumSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 8;
+        long actual = service.calculateMaximumSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -49,9 +48,9 @@ public class StatsServiceTest {
     @Test
     public void shouldCountBelowAverageSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 5;
-        int actual = service.countBelowAverageSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 5;
+        long actual = service.countBelowAverageSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -59,9 +58,9 @@ public class StatsServiceTest {
     @Test
     public void shouldCountAboveAverageSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 5;
-        int actual = service.countAboveAverageSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 5;
+        long actual = service.countAboveAverageSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
